move lib and macros to joinrun, move Chymyst.scala out to chymyst, simplify project build
