+++ .travis.yml
@@ -16,7 +16,7 @@ deploy:
   provider: releases
   api_key:
     secure: "QtQK+MK4yylLKyLV6wDF4uuT9R9i/8OhK7OGC8pKAIzu9wylWGMfFmje3JwD69Ek0yd06h+3xG8OU8rRAfQ+hfCRubXGHKi76zeouvxe8BGpbQ6PSsrnUZV6mmOKFWG2EMhVds+r8vZkV/LcT8+RJ6hZoQWkxrsQ7teiV/Bxm5FYiN3hQJHjbJ5H7jRKBRDnDTI1LbTejpVaNCYq7habYhjSYwHpzfuLrtoE0RZvArAz06ltG8iAgLUZiYGPzv3eVWwhhtsj+dTX+B4p5hvdj5ULeRUdU5+XbBwu6+ZneYI1nDLZCxQiN2LWmqhfBvYSeYsmAOySexueOMIwL/+92mD3EHTNoXs39ZX0DSa8sulLQdq+SCeTc7F2SlpsZVesmDm3Y2cfGC5uhIdj7TeP0XD7ex5sL7Wt7gYrrEjhghxaTbyMZy9KtGHw0YEfSM3n9QbDXPun9mk7YPJs+m1jvTa5cL5kxFWI+Oz3eoFV/tSpmVNfxjTIQLlWcasriotMjIM1JoO/qcAe8eoLefpJnfsZZK3IdBRtPm2Gvv+b4I4ET/td/N5k4cFJqp6++vk6ByE+MnctYsOmeftpNPlLRAfh4Ma3OTg9xbaTXUWV62E39XYb6RRTEj0KfI3wvOounWcYIZdhUnLr8hBkczXO8AsPlOHviYU7fpldut6tCUs="
-  file: "*/target/scala-2.11/*.jar"
+  file: "joinrun/target/scala-2.11/*.jar"
   skip_cleanup: true
   on:
     tags
