+++ joinrun/src/main/scala/code/chymyst/jc/ConjunctiveNormalForm.scala
