+++ joinrun/src/main/scala/code/chymyst/jc/SmartThread.scala
