+++ joinrun/src/test/scala/code/chymyst/jc/Sha1Props.scala
